flaskApp.Folder/Flask.App/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
azure-common==1.1.28
azure-core==1.29.5
azure-functions==1.17.0
azure-servicebus==7.11.4
bidict==0.22.1
blinker==1.7.0
cachelib==0.10.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.7
cryptography==40.0.2
cycler==0.12.1
exceptiongroup==1.2.0
fastapi==0.104.1
Flask==3.0.0
Flask-Session==0.5.0
Flask-SocketIO==5.3.6
Flask-Testing==0.8.1
fonttools==4.46.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.5.1
msal==1.21.0
msal-extensions==1.0.0
msrest==0.7.1
mysql==0.0.3
mysqlclient==2.2.0
numpy==1.26.2
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
portalocker==2.8.2
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
PyMySQL==1.0.3
pyparsing==3.1.1
pytest==7.4.3
python-dateutil==2.8.2
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3.post1
requests==2.27.1
requests-oauthlib==1.3.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
twilio==7.9.2
typing_extensions==4.9.0
tzlocal==5.2
uamqp==1.6.6
urllib3==1.26.18
uvicorn==0.23.2
Werkzeug==3.0.1
wsproto==1.2.0